refactor(portal): extract profile section into ProfileSection.svelte
moves the artist profile form (display name / bio / avatar / support link)
out of the portal page into a co-located ProfileSection.svelte. parent
drops from 1461 -> 987 lines.

the shared atprotofans eligibility flag stays parent-owned (TracksSection
also consumes it); ProfileSection receives it as props and self-loads its
own profile fields on mount.
